handle optional s3fs dependency
