+++ docs/source/package_reference/main_classes.rst
@@ -5,11 +5,11 @@ Main classes
 ~~~~~~~~~~~~~~~~~~~~~
 
 
-.. autoclass:: datasets.filesystem.S3FileSystem(anon=False, key=None, secret=None, token=None, use_ssl=True, client_kwargs=None, requester_pays=False, default_block_size=None, default_fill_cache=True, default_cache_type='bytes', version_aware=False, config_kwargs=None, s3_additional_kwargs=None, session=None, username=None, password=None, asynchronous=False, loop=None, **kwargs)
+.. autoclass:: datasets.filesystems.S3FileSystem(anon=False, key=None, secret=None, token=None, use_ssl=True, client_kwargs=None, requester_pays=False, default_block_size=None, default_fill_cache=True, default_cache_type='bytes', version_aware=False, config_kwargs=None, s3_additional_kwargs=None, session=None, username=None, password=None, asynchronous=False, loop=None, **kwargs)
  
-.. autofunction:: datasets.filesystem.preproc_dataset_path
+.. autofunction:: datasets.filesystems.preproc_dataset_path
 
-.. autofunction:: datasets.filesystem.is_remote_filesystem
+.. autofunction:: datasets.filesystems.is_remote_filesystem
 
 
 ``DatasetInfo``
